+++ CHANGELOG.md
@@ -70,6 +70,8 @@ All notable changes to this project will be documented in this file. See [standa
   }
   ```
 
+- 设置 `packing/config/packing.js` 中的 `templateInjectPosition: false` 来关闭生成网页文件时自动注入 assets 的功能
+
 <a name="2.6.6"></a>
 ## [2.6.6](https://github.com/packingjs/packing/compare/v2.6.5...v2.6.6) (2018-02-28)